Rename text-control to title-input and settings-drawer to settings
- Rename text-control directory to title-input to match component name
- Rename settings-drawer directory to settings to match component name
- Update SettingsDrawer component to Settings
- Update all imports, references, and classNames
- Update Storybook stories to use new component names
- Preserve file history using git mv
