--- packages/block-editor/src/components/link-control-v2/settings-drawer/index.tsx
@@ -1,5 +1,0 @@
-/**
- * Internal dependencies
- */
-export { SettingsDrawer } from './component';
-
