+++ packages/block-editor/src/components/link-control-v2/settings/index.tsx
@@ -1,0 +1,5 @@
+/**
+ * Internal dependencies
+ */
+export { Settings } from './component';
+
